Make changes to enable visualization

+++ rllab/envs/proxy_env.py
@@ -36,7 +36,10 @@ class ProxyEnv(Env, Serializable):
         return self._wrapped_env.horizon
 
     def terminate(self):
-        self._wrapped_env.terminate()
+        if hasattr(self._wrapped_env, "env"):
+            self._wrapped_env.env.terminate()
+        else:
+                self._wrapped_env.terminate()
 
     def get_param_values(self):
         return self._wrapped_env.get_param_values()
